+++ test/cctest/test_base64.cc
@@ -1,4 +1,4 @@
-#include "base64.h"
+#include "base64-inl.h"
 
 #include <cstddef>
 #include <cstring>
